// frontend/src/api/model/sessionWithResultsDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * TeamPrompts.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { GenerationResultDto } from './generationResultDto';
import type { SessionDto } from './sessionDto';

export interface SessionWithResultsDto {
  session: SessionDto;
  results: GenerationResultDto[];
}